backend/requirements.txt
Flask==2.3.3
Flask-MySQLdb==2.0.0
python-dotenv==1.0.1
Werkzeug==2.3.7
Flask-Cors==3.0.10